requirements.txt
asgiref==3.6.0
async-timeout==4.0.3
blinker==1.6.2
Brotli==1.0.9
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.7
colorama==0.4.6
ConfigArgParse==1.7
crispy-bootstrap4==2022.1
cryptography==40.0.2
defusedxml==0.7.1
Django==4.2
django-allauth==0.54.0
django-crispy-forms==2.0
django-markdownx==4.0.2
django-redis==5.3.0
djangorestframework==3.14.0
drf-yasg==1.21.7
environ==1.0
Flask==2.3.3
Flask-BasicAuth==0.2.0
Flask-Cors==4.0.0
gevent==23.7.0
geventhttpclient==2.0.9
greenlet==2.0.2
idna==3.4
inflection==0.5.1
itsdangerous==2.1.2
Jinja2==3.1.2
locust==2.16.1
Markdown==3.4.3
MarkupSafe==2.1.3
msgpack==1.0.5
mysqlclient==2.2.0
oauthlib==3.2.2
packaging==23.1
Pillow==9.5.0
psutil==5.9.5
pycparser==2.21
PyJWT==2.7.0
python-dotenv==1.0.0
python3-openid==3.2.0
pytz==2023.3.post1
pywin32==306
PyYAML==6.0.1
pyzmq==25.1.1
redis==5.0.0
requests==2.31.0
requests-oauthlib==1.3.1
roundrobin==0.0.4
six==1.16.0
sqlparse==0.4.3
typing_extensions==4.7.1
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.2
waitress==2.1.2
Werkzeug==2.3.7
whitenoise==6.4.0
zope.event==5.0
zope.interface==6.0
